changed int to double to suppress compiler warnings

+++ src/board/mark.h
@@ -19,8 +19,8 @@ public:
 	virtual void setPos(double,double) = 0;
 //	virtual void setY(double) = 0;
 	virtual void setSize(double, double) = 0;
-	virtual int getSizeX() const = 0;
-	virtual int getSizeY() const = 0;
+	virtual double getSizeX() const = 0;
+	virtual double getSizeY() const = 0;
 	virtual void show() = 0;
 	virtual void hide() = 0;
 	virtual void setColor(const QColor &col) = 0;
@@ -41,15 +41,15 @@ protected:
 class MarkSquare : public QGraphicsRectItem, public Mark
 {
 public:
-	MarkSquare(int x, int y, int size, QGraphicsScene *canvas, QColor col=Qt::black);
+	MarkSquare(int x, int y, double size, QGraphicsScene *canvas, QColor col=Qt::black);
 	virtual ~MarkSquare() {/* hide(); */}
 	virtual int type() const { return RTTI_MARK_SQUARE; }
 	virtual MarkType getType() const { return markSquare; }
 	virtual void setPos(double x, double y) { QGraphicsRectItem::setPos((qreal)x, (qreal)y); }
 //	virtual void setY(double y) { QGraphicsRectItem::setY(y); }
 	virtual void setSize(double x, double y) { setRect(rect().x()/*myX*/,rect().y()/* myY*/, x*0.5, y*0.5); }
-	virtual int getSizeX() const { return (int) rect().width(); }
-	virtual int getSizeY() const { return (int) rect().height(); }
+	virtual double getSizeX() const { return (int) rect().width(); }
+	virtual double getSizeY() const { return (int) rect().height(); }
 	virtual void show() { QGraphicsRectItem::show(); }
 	virtual void hide() { QGraphicsRectItem::hide(); }
 	virtual void setColor(const QColor &col) { setPen(QPen(col, 1)); }
@@ -59,7 +59,7 @@ public:
 class MarkCircle : public QGraphicsEllipseItem, public Mark
 {
 public:
-	MarkCircle(int x, int y, int size, QGraphicsScene *canvas, QColor col=Qt::black, bool s=false);
+	MarkCircle(int x, int y, double size, QGraphicsScene *canvas, QColor col=Qt::black, bool s=false);
 	virtual ~MarkCircle() { /*hide();*/ }
 	virtual int type() const { return RTTI_MARK_CIRCLE; }
 	virtual MarkType getType() const { return markCircle; }
@@ -67,8 +67,8 @@ public:
 //	virtual void setX(double x) { QGraphicsEllipseItem::setX(x + width()/2.0); }
 //	virtual void setY(double y) { QGraphicsEllipseItem::setY(y + width()/2.0); }
 	virtual void setSize(double x, double y);
-	virtual int getSizeX() const { return (int)rect().width(); }
-	virtual int getSizeY() const { return (int)rect().height(); }
+	virtual double getSizeX() const { return (int)rect().width(); }
+	virtual double getSizeY() const { return (int)rect().height(); }
 	virtual void show() { QGraphicsEllipseItem::show(); }
 	virtual void hide() { QGraphicsEllipseItem::hide(); }
 	virtual void setColor(const QColor &col) { setPen(QPen(col, 1)); }
@@ -84,7 +84,7 @@ protected:
 class MarkTriangle : public QGraphicsPolygonItem, public Mark
 {
 public:
-	MarkTriangle(int x, int y, int size, QGraphicsScene *canvas, QColor col=Qt::black);
+	MarkTriangle(int x, int y, double size, QGraphicsScene *canvas, QColor col=Qt::black);
 	virtual ~MarkTriangle() { /*hide();*/ }
 	virtual int type() const { return RTTI_MARK_TRIANGLE; }
 	virtual MarkType getType() const { return markTriangle; }
@@ -92,8 +92,8 @@ public:
 //	virtual void setX(double x) { QGraphicsPolygonItem::setX(x); }
 //	virtual void setY(double y) { QGraphicsPolygonItem::setY(y + size-1); }
 	virtual void setSize(double, double);
-	virtual int getSizeX() const { return size; }
-	virtual int getSizeY() const { return size; }
+	virtual double getSizeX() const { return size; }
+	virtual double getSizeY() const { return size; }
 	virtual void show() { QGraphicsPolygonItem::show(); }
 	virtual void hide() { QGraphicsPolygonItem::hide(); }
 	virtual void setColor(const QColor &col) { setPen(QPen(col, 1)); }
@@ -103,7 +103,7 @@ protected:
 //	virtual void drawShape(QPainter & p);
 	void paint ( QPainter * painter, const QStyleOptionGraphicsItem * option, QWidget * widget = 0 ) { painter->setRenderHints(QPainter::Antialiasing); QGraphicsPolygonItem::paint(painter, option, widget);}
 private:
-	int size;
+	double size;
 	QPolygon pa;
 };
 
@@ -120,7 +120,7 @@ public:
 class MarkCross : public QGraphicsLineItem, public Mark
 {
 public:
-	MarkCross(int x, int y, int s, QGraphicsScene *canvas, QColor col=Qt::black, bool plus=false);
+	MarkCross(int x, int y, double s, QGraphicsScene *canvas, QColor col=Qt::black, bool plus=false);
 	virtual ~MarkCross();
 	virtual int type() const { return RTTI_MARK_CROSS; }
 	virtual MarkType getType() const { return markSquare; }
@@ -128,8 +128,8 @@ public:
 //	virtual void setX(double x) { QGraphicsLineItem::setX(x); if (ol != NULL) ol->setX(x); }
 //	virtual void setY(double y) { QGraphicsLineItem:setY(y); if (ol != NULL) ol->setY(y); }
 	virtual void setSize(double, double);
-	virtual int getSizeX() const { return size; }
-	virtual int getSizeY() const { return size; }
+	virtual double getSizeX() const { return size; }
+	virtual double getSizeY() const { return size; }
 	virtual void show() { QGraphicsLineItem::show(); }
 	virtual void hide() { QGraphicsLineItem::hide(); }
 	virtual void setColor(const QColor &col);
@@ -137,14 +137,14 @@ public:
 	
 private:
 	MarkOtherLine *ol;
-	int size;
+	double size;
 	bool plussign;
 };
 
 class MarkText : public QGraphicsSimpleTextItem, public Mark
 {
 public:
-	MarkText( int x, int y, int size, const QString &txt, QGraphicsScene *canvas, QColor col=Qt::black,
+	MarkText( int x, int y, double size, const QString &txt, QGraphicsScene *canvas, QColor col=Qt::black,
 		short c=-1, bool bold=false, bool overlay=true);
 	virtual ~MarkText();
 	virtual int type() const { return RTTI_MARK_TEXT; }
@@ -153,8 +153,8 @@ public:
 //	virtual void setX(double x) { QGraphicsRectItem::setX(x); if (rect != NULL) rect->setX(x); }
 //	virtual void setY(double y) { QGraphicsRectItem::setY(y); if (rect != NULL) rect->setY(y); }
 	virtual void setSize(double, double);
-	virtual int getSizeX() const { return width; }
-	virtual int getSizeY() const { return height; }
+	virtual double getSizeX() const { return width; }
+	virtual double getSizeY() const { return height; }
 	virtual void show() { QGraphicsSimpleTextItem::show(); }
 	virtual void hide() { QGraphicsSimpleTextItem::hide(); }
 	virtual void setColor(const QColor &col) { setPen(QPen(col, 1));}//setColor(col); }
@@ -179,7 +179,7 @@ private:
 class MarkNumber : public MarkText
 {
 public:
-	MarkNumber( int x, int y, int size, short num, QGraphicsScene *canvas, QColor col=Qt::black, bool bold=false)
+	MarkNumber( int x, int y, double size, short num, QGraphicsScene *canvas, QColor col=Qt::black, bool bold=false)
 		: MarkText( x, y, size, QString::number(num+1), canvas, col, num, bold)
 	{ }
 	virtual int type() const { return RTTI_MARK_NUMBER; }
@@ -189,7 +189,7 @@ public:
 class MarkTerr : public MarkCross
 {
 public: 
-	MarkTerr(int x, int y, int s, StoneColor c, QGraphicsScene *canvas)
+	MarkTerr(int x, int y, double s, StoneColor c, QGraphicsScene *canvas)
 		: MarkCross(x, y, s, canvas, (c == stoneBlack ? Qt::black :  Qt::white)),
 		col(c)
 	{ }
@@ -203,7 +203,7 @@ private:
 class MarkSmallStoneTerr : public Mark, public QGraphicsPixmapItem
 {
 	public: 
-		MarkSmallStoneTerr(int x, int y, int s, StoneColor c, QList<QPixmap> * p, QGraphicsScene *canvas);
+		MarkSmallStoneTerr(int x, int y, double s, StoneColor c, QList<QPixmap> * p, QGraphicsScene *canvas);
 		virtual ~MarkSmallStoneTerr() {};
 		virtual int type() const { return RTTI_MARK_TERR; }
 		virtual MarkType getType() const { if (col == stoneBlack) return markTerrBlack; else return markTerrWhite; }
@@ -211,8 +211,8 @@ class MarkSmallStoneTerr : public Mark, public QGraphicsPixmapItem
 		int getY() const { return _y; };
 		virtual void setPos(double x, double y) { QGraphicsPixmapItem::setPos((qreal)x , (qreal)y); }
 		virtual void setSize(double, double ) {};
-		virtual int getSizeX() const { return size; }
-		virtual int getSizeY() const { return size; }
+		virtual double getSizeX() const { return size; }
+		virtual double getSizeY() const { return size; }
 		virtual void show() { QGraphicsPixmapItem::show(); };
 		virtual void hide() { QGraphicsPixmapItem::hide(); };
 		virtual void setColor(const QColor &) {};
@@ -220,7 +220,7 @@ class MarkSmallStoneTerr : public Mark, public QGraphicsPixmapItem
 	private:
 		int _x, _y;
 		StoneColor col;
-		int size;
+		double size;
 };
 
 #endif
